optimization: BatchFormValidators module opt
