+++ gcloud/common_template/apis/django/api.py
@@ -30,7 +30,7 @@ from gcloud.template_base.utils import read_template_data_file
 from gcloud.iam_auth.view_interceptors.template import BatchFormInterceptor
 from gcloud.openapi.schema import AnnotationAutoSchema
 from gcloud.template_base.apis.django.api import base_batch_form
-from gcloud.tasktmpl3.apis.django.validators import BatchFormValidator
+from gcloud.template_base.apis.django.validators import BatchFormValidator
 from gcloud.utils.dates import time_now_str
 from gcloud.utils.strings import string_to_boolean
 from gcloud.utils.decorators import request_validate
